Email-free registration
The simplest (and the least 'secure') way to register new users - just
let them in. This should be the default way to do it.  No email setup
needed.  TODO - make a config item so the admin can choose how the users
register. (private request, auto-register with email confirmation, all
in)