Fixed one scrolling issue. Now the messages are auto-scrolled to the top
This one wasn't trivial: AF Framework generates the extra scrollers div
elements with no ID. Their ScrollTop attribute is always 0 and the
scroll method wasn't working. Removing  transform: matrix3d element
seems to do the trick. Hacky way to deal with it for sure; I'd
appreciate the advice on the proper way to scroll the auto-generated
content in AF.